NetworkJumpProcesses.jl

Documentation for NetworkJumpProcesses.jl

Missing docstring.

Missing docstring for ConstantJumpVertex(rate, affect!). Check Documenter's build log for details.

Missing docstring.

Missing docstring for VariableJumpVertex(rate, affect!). Check Documenter's build log for details.

Missing docstring.

Missing docstring for ConstantJumpEdge(rate, affect!). Check Documenter's build log for details.

Missing docstring.

Missing docstring for VariableJumpEdge(rate, affect!). Check Documenter's build log for details.

NetworkJumpProcesses.network_jump_set — Method

network_jump_set(graph; vertex_reactions::Vector{T}=Vector{JumpVertex}(), edge_reactions::Vector{U}=Vector{JumpEdge}()) where {T <: JumpVertex, U <: JumpEdge}

Construct a JumpSet from a Graph and a list of JumpVertex and JumpEdge reactions.

See also: ConstantJumpVertex, ConstantJumpEdge, VariableJumpVertex, VariableJumpEdge

NetworkJumpProcesses.vartojumps — Method

vartojumps(graph, nb_vertex_states, nb_vertex_reacs, nb_edge_reacs)

Create a dependency graph that maps the vertex states to the vertex and edge reactions. This graph can be used for the RSSA and RSSACR aggregators.

See also: jumptovars, Jump Aggregators Requiring Dependency Graphs

NetworkJumpProcesses.jumptovars — Method

jumptovars(graph, nb_vertex_states, nb_vertex_reacs, nb_edge_reacs)

Create a dependency graph that maps the vertex and edge reactions to the vertex states. This graph can be used for the RSSA and RSSACR aggregators.

See also: vartojumps, Jump Aggregators Requiring Dependency Graphs