Reference

Documentation for NetworkJumpProcesses.jl

NetworkJumpProcesses.ConstantJumpEdge — Type

ConstantJumpEdge(rate, affect!)

Construct a constant rate jump over an edge with rate rate and affect! methods.

Arguments

rate::Function: signature (vs, vd, p, t) -> Real

affect!::Function: signature (vs, vd, nghbs, p, t) -> nothing

See also: Types of Jumps

NetworkJumpProcesses.ConstantJumpVertex — Type

ConstantJumpVertex(rate, affect!)

Construct a constant rate jump vertex with rate rate and affect! methods.

Arguments

rate::Function: signature (v, nhgbs, p, t) -> Real

affect!::Function: signature (v, nghbs, p, t)

See also: Types of Jumps

NetworkJumpProcesses.VariableJumpEdge — Type

VariableJumpEdge(rate, affect!)

Construct a variable rate jump over an edge with rate rate and affect! methods.

Arguments

rate::Function: signature (vs, vd, p, t) -> Real

affect!::Function: signature (vs, vd, nghbs, p, t) -> nothing

See also: Types of Jumps

NetworkJumpProcesses.VariableJumpVertex — Type

VariableJumpVertex(rate, affect!)

Construct a variable rate jump vertex with rate rate and affect! methods.

Arguments

rate::Function: signature (v, nhgbs, p, t) -> Real

affect!::Function: signature (v, nghbs, p, t) -> nothing

See also: Types of Jumps

NetworkJumpProcesses.jumptovars — Function

jumptovars(graph, nb_vertex_reacs, nb_edge_reacs, nb_vertex_states=1)

Create a dependency graph that maps the vertex and edge reactions to the vertex states. This graph can be used for the RSSA and RSSACR aggregators.

See also: vartojumps, Jump Aggregators Requiring Dependency Graphs

NetworkJumpProcesses.network_jump_set — Method

network_jump_set(
    graph; vertex_reactions::Vector{T}=Vector{JumpVertex}(),
    edge_reactions::Vector{U}=Vector{JumpEdge}(), nb_states=1
) where {
    T <: Union{JumpVertex, Vector{<:JumpVertex}},
    U <: Union{JumpEdge, Vector{<:JumpEdge}}
}

Construct a JumpSet from a Graph and a list of JumpVertex and JumpEdge reactions. vertex_reactions and edge_reactions can be either an vector of reactions which will all be applied to every vertex and edge respectively. The other option is a vector of vectors of reactions, where the ith vector of reactions will be applied to the ith vertex or edge. These variables may be mixed. Each vertex has nb_states variables associated.

See also: ConstantJumpVertex, ConstantJumpEdge, VariableJumpVertex, VariableJumpEdge

NetworkJumpProcesses.vartojumps — Method

vartojumps(graph::AbstractGraph, nb_vertex_reacs::T, nb_edge_reacs::U, nb_vertex_states=1) where {
    T <: Union{Integer, Vector{<:Integer}},
    U <: Union{Integer, Vector{<:Integer}}
}

Create a dependency graph that maps the vertex states to the vertex and edge reactions. This graph can be used for the RSSA and RSSACR aggregators. If all vertices (edges) have the same number of reactions, then nb_vertex_reacs (nb_edge_reacs) can be an integer, equal to that number. Otherwise, nb_vertex_reacs (nb_edge_reacs) must be a vector of length nv(graph) (ne(graph)) , such that nb_vertex_reacs[v] (nb_edge_reacs[e]) is the number of reactions associated with vertex v (edge e).

See also: jumptovars, Jump Aggregators Requiring Dependency Graphs

NetworkJumpProcesses.vertex_range — Method

vertex_range(n, v)

Returns the range of vertices for a given vertex v, given that the number of states is n.

NetworkJumpProcesses.vertex_to_edges — Method

vertex_to_edges(graph::AbstractGraph)

Create a dictionary that maps each vertex to the edges it is connected to.